# puts all mex calls to the c++ RigidBodyManipulator class through a single mex function interface