--- a/systems/plants/@RigidBodyManipulator/allCollisions.m
+++ b/systems/plants/@RigidBodyManipulator/allCollisions.m
@@ -35,5 +35,5 @@ function [ptsA,ptsB,idxA,idxB] = allCollisions(obj,kinsol)
       'You must call doKinematics with mex enabled');
   end
 
-  [ptsA,ptsB,idxA,idxB] = allCollisionsmex(obj.mex_model_ptr);
+  [ptsA,ptsB,idxA,idxB] = rigidBodyManipulatormex(obj.ALL_COLLISIONS_COMMAND,obj.mex_model_ptr);
 end