--- a/systems/plants/@RigidBodyManipulator/bodyKin.m
+++ b/systems/plants/@RigidBodyManipulator/bodyKin.m
@@ -137,9 +137,9 @@ if (kinsol.mex)
     error('Drake:RigidBodyManipulator:InvalidKinematics','This kinsol is not valid because it was computed via mex, and you are now asking for an evaluation with non-numeric pts.  If you intended to use something like TaylorVar, then you must call doKinematics with use_mex = false');
   end
   if compute_dP
-    [P, dP] = forwardKinPositionGradientmex(obj.mex_model_ptr, m, body_or_frame_ind, 1);
+    [P, dP] = rigidBodyManipulatormex(obj.FORWARD_KIN_POSITION_GRADIENT_COMMAND,obj.mex_model_ptr, m, body_or_frame_ind, 1);
   elseif compute_P
-    P = forwardKinPositionGradientmex(obj.mex_model_ptr, m, body_or_frame_ind, 1);
+    P = rigidBodyManipulatormex(obj.FORWARD_KIN_POSITION_GRADIENT_COMMAND,obj.mex_model_ptr, m, body_or_frame_ind, 1);
   end
 else
   if compute_P