the goal is to:
  - reduce compile time (it was getting ridiculous)
  - reduce the barrier to exposing more interfaces.  we shouldn't feel any pain at exposing any (or all!) of the public methods
  - reduce the memory footprint - now we have only one mex file in memory instead of N, where N had gotten large

i'd eventually like to see a very clean parallel between the matlab and c++ RBM classes.  this is just a baby step.