@RussTedrake 2015-02-16
btw - opinions are welcome!