@psiorx 2015-02-16
I think readability took a hit here, but it does compile faster.  I don't think contactConstraintsmex will merge with this without modification.