@RussTedrake 2015-02-16
i do worry about readability.  but i would argue that the pollution of *mex.cpp files in the class directory (with lots of copy/paste all over the place, as evidenced by a ridiculous number of comments specifying the wrong mex file in the text) was very messy before.  and the calls to the mex interface almost always happen directly inside the matlab method of the same name -- so having a slightly less (and only slightly less) readable one line call to the mex function doesn't seem bad.

The bad thing is the big file for the single mex interface.  but we have that for the main RBM.cpp, too.  
(also bad is the jumble of different interfaces, including some that appear to do identical things, like getCMM and centroidalMomentumMatrix, but we'll clean that all up eventually).