@RussTedrake 2015-02-16
john - might you be able to look to see if the memory footprint of matlab is any smaller (e.g. when running and atlas example before and after) with the new version?  do you notice any runtime speed differences?