@psiorx 2015-02-17
Yep, putting together a comparison now.