@RussTedrake 2015-02-17
taking a look.

On Feb 16, 2015, at 7:25 PM, John Carter <notifications@github.com> wrote:

> @RussTedrake ,
> runAtlasBalancing fails for me with these changes:
> 
> Attempt to reference field of non-structure array.
> 
> Error in QPController (line 199)
> obj.mex_ptr = SharedDataHandle(QPControllermex(0,obj,obj.robot.getMexModelPtr.ptr,getB(obj.robot),r.umin,r.umax,terrain_map_ptr));
> 
> Error in runAtlasBalancing (line 84)
> qp = QPController(r,{},ctrl_data,options);
> 
> —
> Reply to this email directly or view it on GitHub.
>