@RussTedrake 2015-02-17
that's actually fantastic.  i can reproduce this, too (btw - you might get a nicer assertion message if you build w/ (make clean; make BUILD_TYPE=Debug).  i've found a few similar eigen memory issues that have been nascent in the code, this might be one more, just because we're moving memory around slightly differently now.  it could be something i screwed up, but I didn't touch the QP controller. 
On Feb 16, 2015, at 7:27 PM, John Carter <notifications@github.com> wrote:

> This one looks scary. runAtlasRunning segfaults and brings down MATLAB with this inviting message:
> 
> MATLAB: malloc.c:3223: __libc_calloc: Assertion `!mem || ((((mchunkptr)((char)(mem) - 2(sizeof(size_t)))))->size & 0x2) || av == (((((mchunkptr)((char)(mem) - 2(sizeof(size_t)))))->size & 0x4) ? ((heap_info ) ((unsigned long) (((mchunkptr)((char)(mem) - 2*(sizeof(size_t))))) & ~((2 * (4 * 1024 * 1024 * sizeof(long))) - 1)))->ar_ptr : &main_arena)' failed.
> 
> —
> Reply to this email directly or view it on GitHub.
>