@psiorx 2015-02-17
Oh, nice.  We might have discovered a sleeper.  Eigen memory issues are a lot less terrifying when you know they're there.