@RussTedrake 2015-02-17
found it.  :)
On Feb 16, 2015, at 7:39 PM, John Carter <notifications@github.com> wrote:

> Oh, nice. We might have discovered a sleeper. Eigen memory issues are a lot less terrifying when you know they're there.
> 
> —
> Reply to this email directly or view it on GitHub.
>