@psiorx 2015-02-17
I'm still seeing the issues on GNU/Linux after pulling your fix. (Both with runAtlasBalancing error and the runAtlasRunning segfault)