@psiorx 2015-02-17
I think the QPController code is getting mad because all mex_model_ptrs are currently broken.  There's a bug on line 2113 of RigidBodyManipulator.m that checks for the existence of constructModelmex (which doesn't exist anymore).

We should either remove that or point it to rigidBodyManipulatormex now.