@RussTedrake 2015-02-17
good catch.  (I still had an old constructModelmex on my harddrive). confirmed that it was broken and that your change fixed it.  will push it now.

On Feb 16, 2015, at 9:07 PM, John Carter <notifications@github.com> wrote:

> I think the QPController code is getting mad because all mex_model_ptrs are currently broken. There's a bug on line 2113 of RigidBodyManipulator.m that checks for the existence of constructModelmex (which doesn't exist anymore).
> 
> We should either remove that or point it to rigidBodyManipulatormex now.
> 
> —
> Reply to this email directly or view it on GitHub.
>