@RussTedrake 2015-02-17
Well crap.  That's no good.  :-/

Could it be the lookup of the constant variables for the commands?  (e.g. obj.FORWARD_KIN_COMMAND)?  

On Feb 16, 2015, at 10:11 PM, John Carter <notifications@github.com> wrote:

> Performance comparison (All results are averages over 10 samples):
> 
> Before (RobotLocomotion:master)
> Build time: 3 mins 35 seconds
> runAtlasBalancing: 16.69 seconds
> 
> After (RussTedrake:mex_entry)
> Build time: 3 mins 12 seconds
> runAtlasBalancing: 18.50 seconds
> 
> The build is about 23 seconds faster, on average, but runAtlasBalancing is ~10% slower. Is the entire mex codebase getting loaded into memory every time any mex call happens? It's hard to explain those numbers as the overhead of a switch statement.
> 
> —
> Reply to this email directly or view it on GitHub.
>