@psiorx 2015-02-17
I think we need more testing.  Maybe runAtlasBalancing isn't a representative sample.  I'll get some more comprehensive results with the profiler on a desktop tomorrow.