@psiorx 2015-02-17
Just tested this on an Alienware desktop (average of 10 samples):

mex_entry:
13.0738 runAtlasBalancing
31.9024 runAtlasRunning

master:
12.2753  runAtlasBalancing
31.7866  runAtlasRunning

Difference is less significant in runAtlasRunning (probably because mex calls aren't occurring during playback of the trajectory, which takes a while).

runAtlasBalancing is still ~6.5% slower than master